hepsiburadaSociality/requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
Django==3.1
django-cors-headers==3.4.0
django-environ==0.4.5
djangorestframework==3.11.1
idna==2.10
psycopg2==2.8.5
Pygments==2.6.1
pytz==2020.1
requests==2.24.0
soupsieve==2.0.1
sqlparse==0.3.1
urllib3==1.25.10
